+++ musicalgestures/_motionvideo_mp_run.py
@@ -256,7 +256,7 @@ def mg_motion_mp(
 
         if title is None:
             title = os.path.basename(of + fex)
-        # save plot as an MgImage at motion_plot for parent MgVideo
+        # save plot as an MgImage at motion_plot_image for parent MgVideo
         self.motion_plot_image = MgImage(save_analysis(of, self.fps, com, qom, self.width,
                                    self.height, unit, title, target_name_plot=target_name_plot, overwrite=overwrite))